requirements.txt
yfinance
supabase
python-dotenv
anthropic
seaborn
